Added app.js file
