+++ index.js
@@ -1,40 +1,10 @@
-require('dotenv').config();
-const express = require('express');
-const app = express();
-const cors = require('cors');
-const mongoose = require('mongoose');
+const app = require('./app');
+const http = require('http');
 const logger = require('./utils/logger');
 
-const blogSchema = new mongoose.Schema({
-  title: String,
-  author: String,
-  url: String,
-  likes: Number,
-});
-
-const Blog = mongoose.model('Blog', blogSchema);
-
-const mongoUrl = process.env.MONGODB_URI;
-mongoose.connect(mongoUrl);
-
-app.use(cors());
-app.use(express.json());
-
-app.get('/api/blogs', (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs);
-  });
-});
-
-app.post('/api/blogs', (request, response) => {
-  const blog = new Blog(request.body);
-
-  blog.save().then((result) => {
-    response.status(201).json(result);
-  });
-});
+const server = http.createServer(app);
 
 const PORT = process.env.PORT;
-app.listen(PORT, () => {
+server.listen(PORT, () => {
   logger.info(`Server running on port ${PORT}`);
 });
